requirements.txt
pyTelegramBotAPI
opencv-python
torch
torchvision
tqdm
imgaug
requests
pytorchcv
pandas
tensorboard
rmn
google-cloud-dialogflow